Correct types

+++ example.cc
@@ -15,12 +15,12 @@ public:
   void Compute(OpKernelContext* context) override {
     // Get input tensor
     const Tensor& input_tensor = context->input(0);
-    auto input = input_tensor.flat<int32>();
+    auto input = input_tensor.flat<float>();
     // Create an output tensor
     Tensor* output_tensor = NULL;
     OP_REQUIRES_OK(context, context->allocate_output(0, input_tensor.shape(),
                                                      &output_tensor));
-    auto output = output_tensor->flat<int32>();
+    auto output = output_tensor->flat<float>();
     const int N = input.size();
     for (int i = 1; i < N; i++) {
       output(i) = (float)i;
